Refactoring to make it easier to have subcommands.
