+++ cmd/block-generator/daemon_test.go
