+++ cmd/block-generator/docker/Dockerfile-generator
@@ -10,4 +10,4 @@ RUN go build
 RUN ls
 RUN mkdir /config
 RUN cp test_config.yml /config/config.yml
-CMD ["/bin/sh", "-c", "./block-generator -port 11111 -config /config/config.yml"]
+CMD ["/bin/sh", "-c", "./block-generator --port 11111 --config /config/config.yml"]
